server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
google-generativeai
tavily-python
requests
sentence-transformers
transformers
scikit-learn
numpy
trafilatura
httpx
